rename event to project card
